# src/main/java/io/dfjinxin/modules/sys/controller/SysConfigController.java
///**
// * 2019 东方金信
// *
// *
// *
// *
// */
//
//package io.dfjinxin.modules.sys.controller;
//
//
//import io.dfjinxin.common.annotation.SysLog;
//import io.dfjinxin.common.utils.PageUtils;
//import io.dfjinxin.common.utils.R;
//import io.dfjinxin.common.validator.ValidatorUtils;
//import io.dfjinxin.modules.sys.entity.SysConfigEntity;
//import io.dfjinxin.modules.sys.service.SysConfigService;
//import io.dfjinxin.common.annotation.RequiresPermissions;
//import org.springframework.beans.factory.annotation.Autowired;
//import org.springframework.web.bind.annotation.*;
//
//import java.util.Map;
//
///**
// * 系统配置信息
// *
// * @author Mark sunlightcs@gmail.com
// */
//@RestController
//@RequestMapping("/sys/config")
//public class SysConfigController extends AbstractController {
//	@Autowired
//	private SysConfigService sysConfigService;
//
//	/**
//	 * 所有配置列表
//	 */
//	@GetMapping("/list")
//	@RequiresPermissions("sys:config:list")
//	public R list(@RequestParam Map<String, Object> params){
//		PageUtils page = sysConfigService.queryPage(params);
//
//		return R.ok().put("page", page);
//	}
//
//
//	/**
//	 * 配置信息
//	 */
//	@GetMapping("/info/{id}")
//	@RequiresPermissions("sys:config:info")
//	public R info(@PathVariable("id") Long id){
//		SysConfigEntity config = sysConfigService.getById(id);
//
//		return R.ok().put("config", config);
//	}
//
//	/**
//	 * 保存配置
//	 */
//	@SysLog("保存配置")
//	@PostMapping("/save")
//	@RequiresPermissions("sys:config:save")
//	public R save(@RequestBody SysConfigEntity config){
//		ValidatorUtils.validateEntity(config);
//
//		sysConfigService.saveConfig(config);
//
//		return R.ok();
//	}
//
//	/**
//	 * 修改配置
//	 */
//	@SysLog("修改配置")
//	@PostMapping("/update")
//	@RequiresPermissions("sys:config:update")
//	public R update(@RequestBody SysConfigEntity config){
//		ValidatorUtils.validateEntity(config);
//
//		sysConfigService.update(config);
//
//		return R.ok();
//	}
//
//	/**
//	 * 删除配置
//	 */
//	@SysLog("删除配置")
//	@PostMapping("/delete")
//	@RequiresPermissions("sys:config:delete")
//	public R delete(@RequestBody Long[] ids){
//		sysConfigService.deleteBatch(ids);
//
//		return R.ok();
//	}
//
//}